ENERGY and the associated SERVICE tactic is added. This fixes #15
The SERVICE tactic, which will be used to change ALOS to a number of
abstract infrastructure services, is now in place. Only the SYSTEM agent
can create this tactic as part of it's strategy.  For now, the rule set
for the ENERGY driver is a placeholder until the actual numbers are
determined.