Fixes #41, blocking now required when no GUI
This can only be tested with athena_batch. Make a simple test script
that uses SIM:RUN to advance time, specify the block parm first as NO
and then as YES. Verify that the order is rejected when block is NO and
accepted when block is YES. Verify that the GUI will accept block with
either value.

The test suite is updated, too.