SERVICE tactic refactor
The service tactic has been refactored to include a mode variable that
allows more flexibility.  In addition to setting LOS to a specific
amount, LOS can be set to the required level, expected level or to a +/-
percentage of the current actual level.

Note that setting the LOS to the saturation level can be achieved simply
by setting the LOS to 1.0