Flunkies, orderx execute, and setredo
There is no need for a setredo/redoscript mechanism; added documentation
for a simpler _ad hoc_ mechanism to the orderx(n) man page.

The orderx(n) class no longer has a "flunky" instance variable.
Insteady, the execute method passes the flunky to the _execute method,
which can query the flunky and save any data it likes to order-specific
instance variables.