Strategy names can now be used in orders
The full name of blocks, tactics and conditions can now be used in
orders.  The strategy class has a type method that will convert a name
to an ID and return it to the order so the user can use either the bean
ID or the strategy bean name in orders.