Updated PLANT:* orders
Ran across an error in orderx_dialog, there was a "throw" with the wrong
number of args.  The reason for the throw looks like an out and out
error, so I changed it to call "error" instead. If this is wrong, then
just need to pass an appropriate type to "throw".