orderx(n) preserves input parameter values
Fixed a bug in which the orderx(n) "get" and "getdict" methods could
return values modified during _validate and _execute.  The input
parameter values are now saved in an instance variable called parmdict,
which is copied to the parms() array on validation.  In this way, the
input parameter values are always preserved.