Fixes #60: added history table for activity_nga
The service history is already being saved, so it did not need to be
added.
To test:
Create, lock and run a scenario forward making sure that some groups are
doing activities. Issue the following command at the CLI:
> select * from hist_activity_nga
Verify that there is history in the table.
Verify that there is a parmdb parameter called hist.activity and that it
is defaulted to "on" and when set to "off" the hist_activity_nga table
does not get built.