Fixes #79
The default map background is the same, but now the upper right corner
is at 42 lat, 45 lon and the lower right at 38 lat, 51 lon.  This
corresponds to an area near the Caspian Sea.

To test:
* Bring up Athena with out any scenario and verify that the corners of
the default map are correct.
* Import a geotiff and verify that the geo-references are changed to the
correct location of the Earth.
* Go to File->New and discard the scenario
* Verify that the default geo-locations go back to those described
above.
* Do the same by typing "new" at the command line
* Verify that auto generation of scenarios still works.

Still TODO: Allow the user to configure the corners of a
non-georeferenced map image. (ie. a GIF, PNG or JPG)