Fixed #87. Map images can be assigned geo-locations
Use File->Geo-reference Map Image to bring up the dialog that will allow
you to assign geo-location to map images. See the order documentation in
the helps for a description of this order and exercise the order.

Also, it's now possible that you can assign geo-locations that will
cause neighborhood point(s) to be outside the extent of the map. If this
occurs a red message is displayed at the top of the map viewer informing
you of this. Use the MAP:GEOREF order to cause this situation to come
about.

The test suite has also been updated.