New gofer capability
Added a revised gofer capability to athena(n).  So far, the ACTORS gofer
has been implemented using it.  All of the non-GUI code works; the gofer
field code has not yet been updated.

Note that the old mechanism is still in place and available for use.  It
will remain until no code uses it.