Fixes #96, rolemapfield widget is now in athena(n)
The CURSE tactic is the only module that made use of this widget. Create
some CURSEs and use CURSE tactics to exercise that the widget is working
properly. Validation is also changes a bit, so make sure to test sanity
checking as well (ie. roles go away or change role type).