Fixed #104: athena(n) supports "new", "open"
This fix adds the "reset" and "load" operations to athenadb(n), and then
to athena(n), so that an existing scenario object can be reset to its
initial state, and a existing scenario file can be loaded into it,
replacing any data that was there.

In addition, the application's "app new" and "app open" commands have
been updated to make use of these instead of creating a new scenario
object each time.  This allows the scenario object's executive to
persist across "new" and "open".