Fixes #130: MAP:* orders call `app puts`
'app puts' is no longer called.  This means that the GUI no longer
displays a message at the bottom when a new map is loaded, it's not
clear that this is even useful.

The only order that depends on Tk is MAP:IMPORT:FILE so the map module
test suite has been updated to reflect this.