Revised vardiff creation protocol
Every vardiff constructor requires two values, one from s1/t1 and one
from s2/t2; but different classes have different key values.  The
comparison class's "add" method needs to be able to see whether the two
values are identical or not, so as to avoid creating a vardiff when
there is no change at all.  So it makes sense to put the keys after the
two values in the constructor argument lists.