Fixes #150: hist_deploy_ng added
This is primarily for use by OEL and whether the history is saved is
controlled by the hist.deploy parmdb parameter with a default of "on".

Bring up Athena workbench and deploy some force groups to neighborhoods.
Run forward and issue the following command at the command line:

>  select * from hist_deploy_ng

and verify that the history table exists and is populated.
Unlock and verify the table is gone.

Lock again, run forward and save the scenario. Unlock and load the just
saved scenario. Verify the table is present and populated.