Added scratchdir(n)
The arachne application and ancillary tools now use scratchdir(n) to
locate the application working directory.  This is a distinct use-case
from workdir(n); workdir(n) presumes that no one but the app needs to
know where the working directory is, whereas scratchdir(n) presumes that
there is a master application and one or more ancillary applications
sharing access to the working directory.  Both arachne(1) and
athena_log(1) use the scratchdir(n) module.