Fixes #160, minimums now enforced
Setting any of the parameters that are in the computation for effective
force will cause this problem, which doesn't make sense.  The user
should not be setting these parameters to zero.  Therefore, the fix here
is to impose minimum values on the parameters involved in the
computation of effective force.  This minimum value is 0.01.

To test, try setting any of the AAM parameters in the effective force
computation to zero and verify that an appropriate error message is
produced.

Then set any number of them to the minimum value of 0.01 and see that no
bgerror fires.  It doesn't really make much sense to use the minimum
value, so attrition rates may get weird in the event a parm gets set to
it. At the very least, AAM shouldn't fail.