Fixes #164: the ruleset is updated
The CIVCAS ruleset now has HREL and VREL components as laid out in the
rules document.

As of this writing the ruleset has gone in exactly as specified prior to
the email from Rob Crowson. If the ruleset changes, then it will be
updated.  The rules document itself will be committed separately since
our docs do not live on GitHub. Either way, the rules document should be
consulted to verify that the rules are firing as described.

Automated tests are updated and added and all pass.