sanity: Entity link translation
The entity names returned by the sanity checkers were intended to
correspond to the
URLs used by Arachne, not by the Workbench's Detail Browser; and in any
event
it's more useful for sanity reports in the Detail Browser to link to
browser tabs rather
than appserver pages.  This commit adds an EntityLink translator, that
translates
entity names into the appropriate gui:// links.