Initial fix for slave thread errors
Any errors thrown in an -async thread::send are handled by a special
error proc.
I tried to set this error proc, but couldn't even be sure that it was
being called.

The problem is that we want errors in the bgslave to propagate back to
the master
thread, where they should logged and raised as errors, so that the user
is notified
that something's wrong.  For now, I've got it doing that in an "after
idle" handler in
what's probably an unsafe way: you can't be sure that the object will
still be there.
I need to come back to this with a fresh mind.