Added ORG groups
Also added links to other entities in JSON output as appropriate. For
example, if a neighborhood has a controlling actor, a link to that actor
is provided on output.

The standard is to add "_link" to the attribute that contains the
reference, thus the attribute for the link to the controller is
"controller_link".