Fixed the background thread error handling.
Implemented a new mechanism such that errors in specific bgslave
requests are
handled so as to terminate the request, and other errors are logged and
the user is notified.

It is still necessary to either document the `athena::background init`
call, or make it
available through some call that *is* public and documented.