Fixes #230: variable is initialized properly
The problem occurs because AAM's start method, which initializes the
variable, is called by the foreground thread only.  When an instance of
AAM is created in the background and then told to advance, the variable
is undefined leading to the bgerror.

The fix is to have the variable get initialized in the constructor.