Fixes #190, bgerror no longer fires
There's a few things that needed to be done for this order. First,
projtype should not even be a valid option, it was being ignored anyway.
Second, the lat/long coords provided in this bug are invalid.
Specifically llon should be greater than ulon. So this check is being
made now.

Tests:
* Try sending the order specified in this bug and verity that
appropriate error messages are generated.
* When all parms outside of -data are valid a message box should popup
with an appropriate message about the image data being invalid. The
order will not be rejected in this case, but no map image will be
displayed.
* Verify that the helps for the MAP:IMPORT:DATA order are updated.