Fixes #144, nbhoods are on map now
They are also created as 100x100 pixel squares instead of 10x10, which
makes it easier to see the contents, if desired.

It's still possible to create nbhoods that go off the map by specifying
more than can fit in the height of the default image, but it's not
really practical to do so.