Fixes #180, min values enforced
Similar to the problem in #160, the civilian concern parameters can be
set to 0.0, which causes a computed denominator to be 0.0 leading to a
divide by zero.  The same fix is applied here, the civilian concern
parms can not be set below 0.01.

If it is desired to disable civilian casualties set aam.civcas.limit to
0.