Fixes #256, state set to DISABLED on reset
When econ is reset now, the state is set to DISABLED and the UI should
reflect that.

Tests:
Autogen a scenario and enable econ
Send the "new" command and verify econ is disabled
Autogen a scenario, enable econ and save scenario
Send the "new" command  and load saved scenario, verify econ is enabled.
Send the "new" command again, verify econ is disabled.
Autogen a scenario and save it without enabling econ.
Enable econ and then load the just saved scenario and verify econ is
disabled.

You should get the same results whether you use the command line or File
> New to send the "new" command.