Case Controller introduced
This commit is the beginning of pages to display content for specific
Arachne cases.  The case page contains tabs for objects and the sidebar
contains navigation to other cases. Ultimately, this may need to change
if the number of tabs becomes unwieldy.

In this commit, the actors and neighborhood tabs have content, but
there's more to do. There are tabs for CIV, FRC, and ORG groups, but no
content yet. That will be coming soon.