More content for groups by group type
The groups tab for a case has a dropdown menu that allows the user to
switch between group types or view all groups at once.  The IDs have
been turned into URLs that the routes module recognizes, but it just
directs the URL to the top level cases page since there isn't a view yet
for individual scenario objects.

I suspect we will need an ObjectsController (or something) to deal with
that kind of content.

Still to do:
* display of raw JSON returns for lists of objects.
* History variables
* proper sidebar content