Last Tab service and object display implemented
It's annoying to navigate off and back on to a page that has a tab set
to something other than the default because it keeps going back to the
default.

This commit introduces a LastTab service that keeps track of the last
tab selected for a given page and sets that tab in the page when the
controller for that page is initialized.

Basically, when a page is first visited its default tab is registered
with the service with an ID and the default tab set in the service.
Whenever a new tab is clicked the service is updated for that ID with
the tab. When navigating back to the service is queried by that pages
controller for the last tab and that tab is activated.

Also JSON output for individual scenario objects is displayed when the
ID of that object is clicked.