Fixed #273: Web App Case Manage Tab
The case page in the test Web App now has a "Manage" tab on which the
user can lock and unlock the scenario, and advance time.  In addition,
the case's metadata is displayed at the top of the page, and a variety
of bits of cleanup have been done across the page.