Meta data for history variables now produced
Arachne has a smarturl to return history meta data on a case by case
basis.  Rather than rely on adb command to retrieve valid values for
particular keys, they are extracted from the history themselves.  This
way, only the pertinent values are returned and if there are no values
returned, that means there is no history for that variable.