Fixes #293, the compdb has been updated.
For top level, or primary, outputs the compdb parameters have the form
"primary.a.<output>". For inputs to other outputs, the parameters have
the form "<output>.a.<input>".

Examples:
Civilian group mood has a primary output called "mood" so the "a" compdb
parm is "primary.a.mood".

Civilian group satisfaction is an input called "sat" to civilian group
mood so it has an "a" parameter called "mood.a.sat".

Civilian group satisfaction is also a primary output so it has its own
primary output parameter called "primary.a.sat".

All the "a" values are defaulted to 1.0.

As of this commit, there is no interface to edit compdb parameters yet,
so it is sufficient to test this by starting Arachne and doing some
comparisons to exercise the compdb and verify that you can see causality
chains.