Fixes #298: active parameter now checked
This was overlooked when the comparison chaining code was updated a
while ago when the intention was to have the active check done prior to
the creation of a vardiff.

This should be working now. When a type of vardiff is not active, it
should not appear either as a top level output or as an input to some
output in the comparison chain.

Retest.