Fixes #304: security.n.g is primary
It is also an input to other primary vardiffs such as nbsecurity.n and
support.n.a.

To test, create a scenario and run time forward and perform an analysis
using the analysis tab. Verify that security.n.g shows up both as an
output on its own and as an input to nbsecurity.n.

Also, this commit sets the "active" flag for the driversat and
drivervrel outputs to 0.  These will need to remain disabled until URAM
is updated so that they can be computed correctly.

The help for nbsecurity has been updated to show that population.g and
security.n.g are inputs to it.