Fixed #169: Mods for background threads
Athena Workbench and Arachne usually advance time in a background
thread; for Workbench, it's controlled by the app.bgticks prefs
parameter.  However, this background thread wasn't loading mods.  The
background master now tells the slave thread to load and apply relevant
mods, and the slave thread logs the loaded mods.  Mods to be loaded
into the background thread need to be properly tagged with a Tcl
package that is loaded into the background thread, e.g., "athena" or
"projectlib".